LayoutTests/canvas/philip/tests/2d.transformation.scale.negative-expected.txt

< [index] >
2d.transformation.scale.negative
scale() with negative scale factors works
References: 2d.transformation.scale
Actual output:
Expected output:

Passed
